REL-366417 - PR Review - Improved Xml comments on cancellation request. - Due to logging singleton pattern, only 1 SubSystem is defined for both import/export. Eliminated passing in the value and the LoggerOptions.SubSystem property is now properly assigned. - Refactored object logging to LogExtensions and added tests. - Use a Stopwatch to capture elapsed time. - Changed Statistics.DocCount from Int64 to Int32 to avoid unnecessary conversion. Verified this isn't directly exposed and won't break RIP. - Added constructor to VolumeManager and pass in ILog instance. - Simplified constructors and marked Obsolete where necessary. - The new Exporter constructor is expanded to support both an ILog and CancellationTokenSource. This will allow RIP to migrate once and take advantage of export cancellation.

+++ Source/Relativity.DataExchange.Legacy/Process/Statistics.vb
@@ -23,7 +23,7 @@ Namespace kCura.WinEDDS
 		Private _fileWaitTime As Int64 = 0
 		Private _metadataWaitTime As Int64 = 0
 		Private _sqlTime As Int64 = 0
-		Private _docCount As Int64 = 0
+		Private _docCount As Int32 = 0
 		Private _lastAccessed As System.DateTime
 		Private _documentsCreated As Int32 = 0
 		Private _documentsUpdated As Int32 = 0
@@ -145,11 +145,11 @@ Namespace kCura.WinEDDS
 			End Set
 		End Property
 
-		Public Property DocCount() As Int64
+		Public Property DocCount() As Int32
 			Get
 				Return _docCount
 			End Get
-			Set(ByVal value As Int64)
+			Set(ByVal value As Int32)
 				_lastAccessed = System.DateTime.Now
 				_docCount = value
 			End Set
